Migrate LSP config to separate file
