+++ config/nvim/init.lua
@@ -3,6 +3,7 @@ require "user.keymaps"
 require "user.plugins"
 require "user.colorscheme"
 require "user.cmp"
+require "user.lsp"
 
 -- Telescope find file functions
 vim.api.nvim_set_keymap('n', '<leader>ff', [[<cmd>lua require('telescope.builtin').find_files()<CR>]], { noremap = true })
@@ -21,28 +22,6 @@ require('nvim-treesitter.configs').setup {
   highlight = {
     enable = true
   }
-}
-
--- Set up mason (lsp)
-require("mason").setup()
-require('mason-lspconfig').setup()
-
--- Mappings.
--- See `:help vim.diagnostic.*` for documentation on any of the below functions
-local opts = { noremap=true, silent=true }
-vim.keymap.set('n', '[d', vim.diagnostic.goto_prev, opts)
-vim.keymap.set('n', ']d', vim.diagnostic.goto_next, opts)
-
--- Completion
-local capabilities = vim.lsp.protocol.make_client_capabilities()
-capabilities = require('cmp_nvim_lsp').update_capabilities(capabilities)
-
-require("mason-lspconfig").setup_handlers {
-  function (server_name)
-    require("lspconfig")[server_name].setup {
-      capabilities = capabilities,
-    }
-  end,
 }
 
 -- Set up lualine
