+++ makesymlinks.sh
@@ -17,6 +17,7 @@ files="
   config/nvim/lua/user/plugins.lua
   config/nvim/lua/user/colorscheme.lua
   config/nvim/lua/user/cmp.lua
+  config/nvim/lua/user/lsp.lua
   config/starship.toml
   rbenv/default-gems
   nvm/default-packages
